refactor: reorganize shared components and update import paths
