+++ lib/core/shared/service/heptics/heptic_manager.dart
