+++ lib/core/shared/user_interface/appbar/appbar.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:go_router/go_router.dart';
 
-import '../../constants/app_icons.dart';
+import '../../../constants/app_icons.dart';
 
 class DeFiRaiseAppBar extends StatelessWidget implements PreferredSizeWidget {
   final String? title;
