+++ lib/core/shared/user_interface/components/auth_header.dart
@@ -1,7 +1,7 @@
 import 'package:defifundr_mobile/core/design_system/theme_extension/app_theme_extension.dart';
 import 'package:flutter/material.dart';
 
-import '../../constants/size.dart';
+import '../../../constants/size.dart';
 
 class AuthHeader extends StatelessWidget {
   const AuthHeader({super.key, required this.title, required this.subtitle});
