+++ lib/feature/authentication/login/presentation/login_screen/login_screen.dart
@@ -4,7 +4,6 @@ import 'package:defifundr_mobile/core/constants/app_icons.dart';
 import 'package:defifundr_mobile/core/constants/assets.dart';
 import 'package:defifundr_mobile/core/design_system/app_colors/app_colors.dart';
 import 'package:defifundr_mobile/core/design_system/theme_extension/app_theme_extension.dart';
-import 'package:defifundr_mobile/core/enums/app_text_field_enums.dart';
 import 'package:defifundr_mobile/core/routers/routes_constant.dart';
 import 'package:defifundr_mobile/core/shared/user_interface/appbar/appbar_header.dart';
 import 'package:defifundr_mobile/core/shared/user_interface/components/or_widget.dart';
