+++ lib/feature/finance/presentation/receive_screen.dart
@@ -6,7 +6,7 @@ import 'package:defifundr_mobile/feature/finance/presentation/finance_home_scree
 import 'package:defifundr_mobile/feature/finance/presentation/select_network_screen.dart'; // For Network model (needed for navigation)
 import 'package:go_router/go_router.dart';
 import 'package:defifundr_mobile/core/routers/routes_constant.dart';
-import 'package:defifundr_mobile/core/shared/appbar/appbar.dart'; // Import DeFiRaiseAppBar
+import 'package:defifundr_mobile/core/shared/user_interface/appbar/appbar.dart'; // Import DeFiRaiseAppBar
 
 class ReceiveScreen extends StatefulWidget {
   const ReceiveScreen({Key? key}) : super(key: key);
