+++ lib/feature/onboarding/presentation/individual_account_flow/widgets/country_selection.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:defifundr_mobile/core/shared/textfield/app_text_field.dart';
+import 'package:defifundr_mobile/core/shared/user_interface/textfield/app_text_field.dart';
 
 class CountrySelectionScreen extends StatefulWidget {
   const CountrySelectionScreen({super.key});
