+++ lib/feature/onboarding/presentation/multi_factor_authentication_screen/screens/enable_face_id_screen.dart
@@ -1,8 +1,8 @@
 import 'package:defifundr_mobile/core/design_system/theme_extension/app_theme_extension.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
-import '../../../../../core/shared/buttons/primary_button.dart';
-import '../../../../../core/shared/buttons/secondary_button.dart';
+import '../../../../../core/shared/user_interface/buttons/primary_button.dart';
+import '../../../../../core/shared/user_interface/buttons/secondary_button.dart';
 
 class EnableFaceIdScreen extends StatelessWidget {
   const EnableFaceIdScreen({super.key});
