+++ lib/feature/onboarding/presentation/multi_factor_authentication_screen/screens/pin_created_screen.dart
@@ -1,7 +1,7 @@
 import 'package:defifundr_mobile/core/design_system/theme_extension/app_theme_extension.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
-import '../../../../../core/shared/buttons/primary_button.dart';
+import '../../../../../core/shared/user_interface/buttons/primary_button.dart';
 
 class PinCreatedScreen extends StatelessWidget {
   const PinCreatedScreen({super.key});
